- changing type of column to uint32 since Kusto doesn't like uint64.
 - Adding script to delete a stream set.

+++ tools/Telemetry/helpers/ScriptContextTelemetryProcessor.cs
@@ -63,7 +63,7 @@ namespace Chakra.Utils
                 }
             }
 
-            output_schema.Add(new ColumnInfo("scriptContextLifeSpanMicros", ColumnDataType.ULong));
+            output_schema.Add(new ColumnInfo("scriptContextLifeSpanMicros", ColumnDataType.UInt));
             output_schema.Add(new ColumnInfo("propType", ColumnDataType.String));
             output_schema.Add(new ColumnInfo("propName", ColumnDataType.String));
             output_schema.Add(new ColumnInfo("propValue", ColumnDataType.UInt));
